--- src/main/resources/org/jenkinsci/plugins/periodicbackup/ExcludeByPattern/config.jelly
@@ -1,15 +1,0 @@
-<!--
-  ExcludeByPattern config page
--->
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
-         xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-
-    <f:entry field="excludeByPatternDescription">
-        <!-- This entry is only for help file binding -->
-    </f:entry>
-
-    <f:entry title="${%excludesString.title}" field="excludesString">
-        <f:textbox/>
-    </f:entry>
-
-</j:jelly>
