--- src/main/resources/org/jenkinsci/plugins/periodicbackup/ExcludeByPattern/help-excludeByPatternDescription.html
@@ -1,5 +1,0 @@
-<div>
-    ExcludeByPattern will choose all the files in the Jenkins homedir during the backup except the files that match the excludes list.
-    During the restore it will delete all the deletable files in the Jenkins homedir
-    and then it will write with files in the selected backup.
-</div>
