+++ src/main/resources/org/jenkinsci/plugins/periodicbackup/FullBackup/config.jelly
@@ -32,5 +32,9 @@
         <!-- This entry is only for help file binding -->
     </f:entry>
 
+    <f:entry title="${%excludesString.title}" field="excludesString">
+        <f:textbox/>
+    </f:entry>
+
 
 </j:jelly>
