+++ src/main/resources/org/jenkinsci/plugins/periodicbackup/FullBackup/config.properties
