+++ src/test/java/org/jenkinsci/plugins/periodicbackup/LocalDirectoryTest.java
@@ -24,7 +24,7 @@ public class LocalDirectoryTest extends HudsonTestCase {
         LocalDirectory localDirectory = new LocalDirectory(path, true);
         Date expectedDate = new Date(123);
         BackupObject expected = new BackupObject(
-                new FullBackup(),
+                new FullBackup(""),
                 new ZipStorage(false, 0),
                 new LocalDirectory(new File("c:\\Temp"), true),
                 expectedDate);
@@ -64,7 +64,7 @@ public class LocalDirectoryTest extends HudsonTestCase {
         File tempDirectory = new File(Thread.currentThread().getContextClassLoader().getResource("data/temp").getFile());
         Date testDate = new Date(123-(Calendar.getInstance().get(Calendar.ZONE_OFFSET)));
         LocalDirectory localDirectory = new LocalDirectory(sourceDir, true);
-        BackupObject backupObject = new BackupObject(new FullBackup(), new ZipStorage(false, 0), localDirectory, testDate);
+        BackupObject backupObject = new BackupObject(new FullBackup(""), new ZipStorage(false, 0), localDirectory, testDate);
 
         Iterable<File> result = localDirectory.retrieveBackupFromLocation(backupObject, tempDirectory);
         File expectedResult = new File(tempDirectory, Util.createFileName(Util.generateFileNameBase(testDate), "zip"));
